dev-requirements.txt
# 
# Please do not manually update this file since the requirements are managed
# by pipenv through Pipfile and Pipfile.lock . 
#
# This file is created and managed automatically by the template and it is left
# here only for backwards compatibility reasons with python's ecosystem.
#
# Please use Pipfile to update the requirements.
#
betamax~=0.8.1
coloredlogs~=10.0
coverage~=4.5.1
emoji~=0.5.1
gitwrapperlib~=0.7.0
nose~=1.3.7
nose-htmloutput~=0.6.0
prospector~=1.1.3
semver~=2.8.1
sphinx~=1.7.9
tox~=3.2.1
twine~=1.11.0